Update code to handle more complicated scenarios
Update code to select the entry for arrays based on the browser version in the array. This handles prefix support upto a certain version and full support after that. Previously, the data would be not supported until fully supported.

Handle version_added being a range value or true.

Handle version_removed having a value.